+++ applications/precision/numeric/residual.cpp
@@ -17,6 +17,7 @@
 
 #include <universal/number/areal/areal.hpp>
 #include <universal/number/posit/posit.hpp>
+#include <universal/number/posit/quire.hpp>
 #include <universal/number/lns/lns.hpp>
 
 // Stillwater BLAS library