+++ applications/reproducibility/blas/l1_fused_dot.cpp
@@ -29,6 +29,7 @@
 // enable posit arithmetic exceptions
 #define POSIT_THROW_ARITHMETIC_EXCEPTION 1
 #include <universal/number/posit/posit.hpp>
+#include <universal/number/posit/fdp.hpp>
 #include <blas/blas.hpp>
 
 template<typename Vector>