+++ applications/reproducibility/blas/l3_fused_mm.cpp
@@ -12,6 +12,7 @@
 // enable posit arithmetic exceptions
 #define POSIT_THROW_ARITHMETIC_EXCEPTION 1
 #include <universal/number/posit/posit.hpp>
+#include <universal/number/posit/fdp.hpp>
 #define BLAS_TRACE_ROUNDING_EVENTS 1
 #include <blas/blas.hpp>
 #include <blas/ext/posit_fused_blas.hpp>   // addition of fdp, fmv, and fmm functions