+++ applications/reproducibility/blas/norms.cpp
@@ -29,6 +29,7 @@
 // enable posit arithmetic exceptions
 #define POSIT_THROW_ARITHMETIC_EXCEPTION 1
 #include <universal/number/posit/posit.hpp>
+#include <universal/number/posit/quire.hpp>
 // Stillwater BLAS library
 #include <blas/blas.hpp>