+++ education/number/posit/exceptions.cpp
@@ -9,6 +9,7 @@
 // enable/disable posit arithmetic exceptions
 #define POSIT_THROW_ARITHMETIC_EXCEPTION 1
 #include <universal/number/posit/posit.hpp>
+#include <universal/number/posit/quire.hpp>
 
 int main()
 try {