+++ education/number/posit/signalling_nar.cpp
@@ -10,6 +10,7 @@
 // disable posit arithmetic exceptions
 #define POSIT_THROW_ARITHMETIC_EXCEPTION 0
 #include <universal/number/posit/posit.hpp>
+#include <universal/number/posit/quire.hpp>
 
 int main()
 try {