+++ education/quire/quires.cpp
@@ -9,6 +9,7 @@
 #define POSIT_THROW_ARITHMETIC_EXCEPTION 0
 // type definitions for the important types, posit<> and quire<>
 #include <universal/number/posit/posit.hpp>
+#include <universal/number/posit/quire.hpp>
 #include <universal/verification/quire_test_suite.hpp>
 
 #define MANUAL_TESTING 1