+++ include/sw/universal/number/posit/fdp.hpp
@@ -8,6 +8,7 @@
 #include <iostream>
 #include <vector>
 #include <universal/traits/posit_traits.hpp>
+#include <universal/number/posit/quire.hpp>
 
 namespace sw { namespace universal {