+++ include/sw/universal/number/posit/posit_fwd.hpp
@@ -1,5 +1,5 @@
 #pragma once
-// posit_fwd.hpp :  forward declarations of the posit/quire environment
+// posit_fwd.hpp :  forward declarations of the posit environment
 //
 // Copyright (C) 2017 Stillwater Supercomputing, Inc.
 // SPDX-License-Identifier: MIT
@@ -10,11 +10,6 @@
 #include <universal/internal/blocktriple/blocktriple.hpp>
 
 namespace sw { namespace universal {
-
-	namespace internal {
-		// generalized floating point type
-		template<unsigned fbits> class value;
-	}
 
 // posit types
 template<unsigned nbits, unsigned es, typename bt> class posit;
@@ -28,12 +23,4 @@ template<unsigned nbits, unsigned es, typename bt, unsigned fbits, BlockTripleOp
 
 template<unsigned nbits, typename bt> int decode_regime(const blockbinary<nbits, bt>&);
 
-// quire types
-template<unsigned nbits, unsigned es, unsigned capacity> class quire;
-template<unsigned nbits, unsigned es, typename bt> internal::value<2*(nbits-2-es)> quire_mul(const posit<nbits,es,bt>&, const posit<nbits,es,bt>&);
-
-// bridge: convert internal::value to posit (needed by quire output path)
-template<unsigned nbits, unsigned es, typename bt, unsigned fbits> posit<nbits,es,bt>& convert(const internal::value<fbits>&, posit<nbits,es,bt>&);
-
 }} // namespace sw::universal
-