+++ linalg/blas/matrix_ops.cpp
@@ -8,6 +8,7 @@
 #include <limits>
 // pull in the number systems you would like to use
 #include <universal/number/posit/posit.hpp>
+#include <universal/number/posit/fdp.hpp>
 #include <universal/number/cfloat/cfloat.hpp>
 #include <universal/number/integer/integer.hpp>