+++ tools/cmd/propp.cpp
@@ -5,6 +5,7 @@
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <universal/number/posit/posit.hpp>
+#include <universal/number/posit/quire.hpp>
 
 typedef std::numeric_limits< double > dbl;
 const char* msg = "arithmetic properties of a posit<16, 1> environment\n\